# docs/swagger.yaml
definitions:
...
  dto.PageResponse:
    properties:
      count:
        description: Total number of items that match the query
        type: integer
      next:
        description: URL to the next page of results
        type: string
      num_pages:
        description: Total number of pages that match the query
        type: integer
      per_page:
        description: Total number of items per page that match the query
        type: integer
      previous:
        description: URL to the previous page of results
        type: string
    type: object
...
  dto.SaltReturnPageResponse:
    properties:
      paging:
        $ref: '#/definitions/dto.PageResponse'
      results:
        description: Array of salt return results
        items:
          $ref: '#/definitions/model.SaltReturn'
        type: array
    type: object
  github_com_jackc_pgtype.JSON:
    properties:
      bytes:
        items:
          type: integer
        type: array
      status:
        $ref: '#/definitions/github_com_jackc_pgtype.Status'
    type: object
...
  github_com_jackc_pgtype.Status:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - Undefined
    - "Null"
    - Present
...
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
...
  model.SaltReturn:
    properties:
      alterTime:
        type: string
      fullRet:
        $ref: '#/definitions/github_com_jackc_pgtype.JSON'
      fun:
        type: string
      id:
        type: string
      jid:
        type: string
      return:
        $ref: '#/definitions/github_com_jackc_pgtype.JSON'
      success:
        type: string
      uuid:
        type: string
    type: object
...
info:
  contact:
    email: pmartin@gatech.edu
    name: API Support
    url: http://oit.gatech.edu
  description: This the Agartha API Backend
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Agartha API
  version: "1.0"
paths:
...
  /api/v1/salt_return:
    get:
      consumes:
      - application/json
      description: Get all SaltReturns
      parameters:
      - description: Filter items by minion id (Supports wildcards * and ? for single
          char matches.)
        in: query
        name: minion_id
        type: string
      - description: Filter items by function (Supports wildcards * and ? for single
          char matches.)
        in: query
        name: fun
        type: string
      - description: Filter items by success status (true/false). This can be null
          so it's a string in the database.
        in: query
        name: success
        type: string
      - description: Filter items from this date (RFC3339 format).
        in: query
        name: since
        type: string
      - description: Filter items up to this date (RFC3339 format).
        in: query
        name: until
        type: string
      - description: restrict to X results
        in: query
        name: per_page
        type: integer
      - description: Page number of results to retrieve
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SaltReturnPageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - Bearer: []
      tags:
      - SaltReturn
  /api/v1/salt_return/{jid}:
    get:
      consumes:
      - application/json
      description: Get all SaltReturns for a specific jid
      parameters:
      - description: jid of the salt return item to retrieve
        in: path
        name: jid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SaltReturn'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - Bearer: []
      tags:
      - SaltReturn
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"